OneDrive - Accenture/GenAI/Similarity-search-tool-cloud-deploy/requirements.txt
Flask==2.2.2
Werkzeug==2.3.7
gunicorn==20.1.0
google-cloud-aiplatform